feat: unified QR code scanner across the app
